Update editing.py
Combined the 3 main functions into one by passing in the main file path as a parameter, as well as added logic to handle the scanning.json file slightly different, since we do not need to "move" objects, just edit or delete.
Added some pretty print formatting to the editing menu... makes it easier to look at the item you are editing and see if it's the one you want to edit.
Finally, created a display_data function that combines the 3 different (but oh-so-similar) data displays. This uses similar logic to determine which headers to display.